Initial refactoring of InMemoryFunctionCatalog
interim

Fix FunctionalExporterTests

final
